--- src/kb/mod.rs
@@ -1,5 +1,0 @@
-// TODO: Move knowledge base.rs into this file
-// TODO: This would make the path less confusing kb::KnowledgeBase, rather than kb::knowledge_base::KnowledgeBase
-pub mod knowledge_base;
-mod parser;
-mod symbols;
