+++ src/main.rs
@@ -1,30 +1,10 @@
-extern crate nom;
-extern crate weak_table;
+extern crate rust_kb;
 
-mod kb;
-
-use kb::knowledge_base::*;
+use rust_kb::KnowledgeBase;
 
 fn main() {
-    //let mut kb = KnowledgeBase::from_file("test/full.kb").unwrap();
-
-//    let f = Fact::new(
-//        kb.intern_string("isa"),
-//        vec![kb.intern_string("?a"), kb.intern_string("?b")],
-//        vec![],
-//    );
-//    let a = kb.query(&f);
-//
-//    for binding in a.iter() {
-//        for symbol_bind in binding.iter() {
-//            print!(
-//                "{} : {}, ",
-//                (*symbol_bind.0).to_string(),
-//                (*symbol_bind.1).to_string()
-//            );
-//        }
-//        println!("");
     let mut kb = KnowledgeBase::new();
+
     if let Ok(fact) = kb.create_fact("fact: (isa box hello);") {
         match kb.assert(fact) {
             Ok(_) => {},
@@ -36,17 +16,4 @@ fn main() {
     if let Ok(fact) = kb.create_fact("fact: (isa box hello);") {
         println!("{:?}",kb.ask(&fact));
     }
-
-
-
-
-    //let f = Fact::new(kb.intern_string("isa"), vec![kb.intern_string("?a"), kb.intern_string("?b")]);
-    //let a = kb.query(&f);
-
-//    for binding in a.iter() {
-//        for symbol_bind in binding.iter() {
-//            print!("{} : {}, ",(*symbol_bind.0).to_string(),(*symbol_bind.1).to_string());
-//        }
-//        println!("");
-//    }
 }
