+++ src/symbols.rs
